Updated HLS Implementation

+++ Vivado/acc_b.prj/acc_b.c
@@ -1,7 +1,7 @@
 #include "acc_b.h"
 
-int acc_b(int b_in, bool b_en ) {
-	static int cnt = 0;
+float acc_b(float b_in, bool b_en ) {
+	static float cnt = 0;
 	if(!b_en) {
 		cnt += b_in;
 	}
